pandaproxy: Invert kafka::error_code dependency
Move `make_error_code`  and `error_category` from pandaproxy to kafka.
